🚀 [Feature]: Initial version (#13)
## Description

This pull request includes several changes to add new functionality for
managing Google Fonts, including fetching, installing, and providing
argument completion for font names and variants. Additionally, it
introduces a new enum for scope management and removes an obsolete
function.

- Fixes #12

### New functionality for Google Fonts:

*
[`scripts/createFontData.ps1`](diffhunk://#diff-e6373952ad8d74d61bb6531ebd6815fcd7a0e4365ddcb4817606d15932fb841eR1-R18):
Added a script to fetch Google Fonts data from the Google Fonts
Developer API and save it to a local file.
*
[`src/functions/public/functions/public/Get-GoogleFont.ps1`](diffhunk://#diff-4bd5ee7ef58bbef5ff0dde3f305d591cf2c628871cda141efba78c26f466d973R1-R42):
Introduced a new function to retrieve Google Fonts assets by name and
variant, with support for wildcards.
*
[`src/functions/public/functions/public/Install-GoogleFont.ps1`](diffhunk://#diff-6e1752dc25f937eb34d8a0a569afa0ae8bee012bd73ba28830b0780f7e995330R1-R109):
Added a function to install Google Fonts to the system, with options for
specifying font names, variants, and installation scope.
*
[`src/functions/public/functions/public/completers.ps1`](diffhunk://#diff-3a17567f5a1709e7b39be318cb407a12121f33cd7a2f1bc08a723c44ce87aff5R1-R16):
Registered argument completers for the `Get-GoogleFont` and
`Install-GoogleFont` functions to provide auto-completion for font names
and variants.

### Codebase simplification:

*
[`src/functions/public/Get-PSModule.ps1`](diffhunk://#diff-40020a52fbe605d8f0d3cc65e32d8df775ebcbcd16c46fefac1051045142d69cL1-L18):
Removed the obsolete `Get-PSModule` function.

### Scope management:

*
[`src/functions/public/classes/private/Scope.ps1`](diffhunk://#diff-34ba5ffd5c4853ca600e8f637052ee3e45d59bdccb932e3a5bc0b74f8254d33bR1-R4):
Added a new `Scope` enum to manage installation scopes for Google Fonts.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas