🩹 [Patch]: Add Update workflow and align with Fonts 1.1.13 (#21)
## Description

This pull request includes several updates to the PowerShell scripts for
managing Google Fonts. The changes primarily focus on improving the
functionality and usability of the `Install-GoogleFont` function,
updating file paths, and refining the descriptions and parameters in the
scripts.

### Functional Improvements:
*
[`src/functions/public/Install-GoogleFont.ps1`](diffhunk://#diff-f926b03377b12f05eed72cf360452ff4ac9b8f4a2c7c858e02068ea43cc050a5L28-R60):
Enhanced the `Install-GoogleFont` function by adding support for
wildcard patterns in font names, enabling pipeline input, and
introducing a `Force` parameter to overwrite existing fonts.
[[1]](diffhunk://#diff-f926b03377b12f05eed72cf360452ff4ac9b8f4a2c7c858e02068ea43cc050a5L28-R60)
[[2]](diffhunk://#diff-f926b03377b12f05eed72cf360452ff4ac9b8f4a2c7c858e02068ea43cc050a5L57-R71)
[[3]](diffhunk://#diff-f926b03377b12f05eed72cf360452ff4ac9b8f4a2c7c858e02068ea43cc050a5L69-R107)

### Path and Data Handling:
*
[`scripts/Update-FontsData.ps1`](diffhunk://#diff-69d235ffacf154789a70d5e00a557ee83393c256cd9431a8b7be57cc41c3ad89L16-R19):
Changed the output path for the `GoogleFonts.json` file to
`src/FontsData.json` and ensured the parent folder is correctly
identified.
*
[`src/variables/private/GoogleFonts.ps1`](diffhunk://#diff-5a49f8240531cc9b0d8a3af80dbf291f01cb92da63a20808e62be252b42623aeL1-R1):
Updated the script to read the Google Fonts data from the local
`FontsData.json` file instead of fetching it from a remote URL.

### Documentation and Descriptions:
*
[`src/functions/public/Get-GoogleFont.ps1`](diffhunk://#diff-86091f30b4dfb1604e5f505f18809aad419502a27c58272d2eb24defcacf6b37L4-R7):
Refined the function's synopsis and example descriptions to remove the
term "asset" and added an `OutputType` attribute.
[[1]](diffhunk://#diff-86091f30b4dfb1604e5f505f18809aad419502a27c58272d2eb24defcacf6b37L4-R7)
[[2]](diffhunk://#diff-86091f30b4dfb1604e5f505f18809aad419502a27c58272d2eb24defcacf6b37L17-R25)

### Module Requirements:
*
[`src/functions/public/Install-GoogleFont.ps1`](diffhunk://#diff-f926b03377b12f05eed72cf360452ff4ac9b8f4a2c7c858e02068ea43cc050a5L1-R1):
Updated the required version of the `Fonts` module to `1.1.13`.

### Minor Changes:
*
[`src/classes/private/Scope.ps1`](diffhunk://#diff-9f8b649c9594a7c7582dae3baf1841e8ffa82e1cbb7e1fae8b560b56f141e6f1L1-R1):
Fixed a minor encoding issue in the `Scope` enum definition.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas