🩹 [Patch]: Auto update of FontsData.json (#23)w
## Description

This pull request includes significant updates to the `Update-FontsData`
workflow and script. The changes aim to automate the process of updating
font data from Google Fonts and creating a pull request with the
updates.

Improvements to workflow:

*
[`.github/workflows/Update-FontsData.yml`](diffhunk://#diff-5c6d6ac6744e613c22fe7a4b5bf6b9fbd4b3bf5548239191e4400a4b17440826L4-R10):
Changed the trigger from `push` to `workflow_dispatch` and `schedule`,
and updated permissions to include `pull-requests: write`.
*
[`.github/workflows/Update-FontsData.yml`](diffhunk://#diff-5c6d6ac6744e613c22fe7a4b5bf6b9fbd4b3bf5548239191e4400a4b17440826L24-R25):
Corrected the script path in the `jobs` section.

Enhancements to the update script:

* Added steps to check out the main branch, pull the latest changes,
create a new branch, retrieve the latest font data, and write the
results to `FontsData.json`.
* Added logic to check for changes, commit and push updates, and create
a pull request using GitHub CLI if there are new changes.

Updates to font data:

* Removed font entries to test the update process.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas